Fix problems in docker-compose-uat.yml
Big problem was that connection string, passed in as environment
variable to mvc container, should not have been surrounded in quotes.
This fixed everything

Other problem is that default of 1GB of memory for each contianer
was not enough. I raised them to 4GB each and suddenly the app
responds immediately instead of intermittenly restarting and
responding with a delay.